# swagger.yaml
openapi: 3.0.3
info:
  title: CargoFlow Logistics API
  version: 2.0.0
  description: >
    API documentation for Shipment Management System.

    USER STORIES & API MAPPING:

    Auth:
    - Register → Create account + Email notification
    - Login → Access system
    - Profile → View/update user details

    Admin:
    - Dashboard → Monitor system
    - Shipments → Create & assign shipments
    - Agents → Manage delivery agents
    - Clients → Manage business clients
    - Tracking → Monitor agent locations

    Client:
    - Shipments → Create & view shipments
    - Tracking → Track deliveries
    - Business → Manage business profile

    Agent:
    - Dashboard → View assigned shipments
    - Status → Update shipment status
    - Location → Update live location
    - Duty → Update availability

    Chatbot:
    - Chat → Ask queries
    - Rebuild → Refresh knowledge base
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register User
      description: "User Story: As a new user, I want to register an account so that I can access the system and receive a confirmation email."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        "201": { description: User registered }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login User
      description: "User Story: As a registered user, I want to log in securely so that I can access my dashboard."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        "200": { description: Login success }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get Profile
      description: "User Story: As a user, I want to view my profile so that I can verify my personal details."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Profile data }

    patch:
      tags: [Auth]
      summary: Update Profile
      description: "User Story: As a user, I want to update my profile so that I can keep my information up to date."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200": { description: Profile updated }

  # ---------------- ADMIN ----------------
...
  /api/v1/agent/dashboard:
    get:
      tags: [Agent]
      summary: Agent Dashboard
      description: "User Story: As a delivery agent, I want to view my assigned shipments so that I can manage my deliveries."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Dashboard }
  /api/v1/agent/shipments/{id}/status:
    patch:
      tags: [Agent]
      summary: Update Shipment Status
      description: "User Story: As a delivery agent, I want to update shipment status so that customers and admins are informed."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateShipmentStatus'
      responses:
        "200": { description: Updated }
...
  /api/v1/agent/update/duty-status:
    patch:
      tags: [Agent]
      summary: Update Duty Status
      description: "User Story: As a delivery agent, I want to update my duty status so that the system knows whether I am available."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DutyStatusUpdate'
      responses:
        "200": { description: Updated }

  # ---------------- CHATBOT ----------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    UserRegister:
      type: object
      required: [name, email, password, phone, city]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        phone: { type: string }
        city: { type: string }
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        password: { type: string }
...
    AgentUpdateShipmentStatus:
      type: object
      properties:
        status: { type: string }
        remarks: { type: string }
        payment_method: { type: string }
    DutyStatusUpdate:
      type: object
      properties:
        status: { type: string }
...